fix some 64bit warnings
git-svn-id: https://svn.code.sf.net/p/pinmame/code/trunk@3786 6f51dfdd-75df-4273-8f3b-818b1e6aa8bf

+++ src/windows/fileio.c
@@ -228,7 +228,7 @@ static char *copy_and_expand_variables(const char *path, int len)
 {
 	char *dst, *result;
 	const char *src;
-	int length = 0;
+	size_t length = 0;
 
 	/* first determine the length of the expanded string */
 	for (src = path; src < path + len; )
